# api-integration/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftAiResponse": {
        "description": "T14b: 单字段 AI 重新生成响应（只读结果，前端决定 PATCH 保存）。",
        "properties": {
          "field": {
            "description": "title/description/attributes/tags",
            "title": "Field",
            "type": "string"
          },
          "value": {
            "description": "俄语 RU 值（非空，无中文/拉丁残留）",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "field",
          "value"
        ],
        "title": "DraftAiResponse",
        "type": "object"
      },
      "DraftAssembleResponse": {
        "description": "POST /drafts/{id}/assemble 响应（v0.70 一键预组装：整卡生成并写回 payload）。",
        "examples": [
          {
            "assembled": [
              "title",
              "description",
              "attributes",
              "tags"
            ],
            "estimated_pricing": {
              "old_price": 910.0,
              "price": 729.0,
              "promo_price": 547.0
            },
            "skipped": [],
            "suggested_category": {
              "category_name": "Автопоилка для животных",
              "description_category_id": "17029651",
              "type_id": "91633"
            },
            "version": 2
          }
        ],
        "properties": {
          "assembled": {
            "description": "实际生成成功并写回 payload 的字段（title/description/attributes/tags 子集）",
            "items": {
              "type": "string"
            },
            "title": "Assembled",
            "type": "array"
          },
          "estimated_pricing": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "三档预估价 RUB {price, old_price, promo_price?}（展示字段；pricing_node 永远按成本+margin 重算，管线不消费）",
            "title": "Estimated Pricing"
          },
          "skipped": {
            "description": "跳过的字段（已含西里尔幂等跳过 / 源为空 / ozon_attributes 已有内容不混源 / 生成失败）",
            "items": {
              "type": "string"
            },
            "title": "Skipped",
            "type": "array"
          },
          "suggested_category": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "类目建议 {description_category_id, type_id, category_name}（展示字段；只写 draft.suggested_category，绝不写 draft.ozon_category 劫持管线仲裁链）",
            "title": "Suggested Category"
          },
          "version": {
            "description": "写回后的草稿版本（version++）",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "DraftAssembleResponse",
        "type": "object"
      },
      "DraftOut": {
        "description": "草稿详情（payload = envelope，不含任何凭证）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "创建时间",
            "title": "Created At"
          },
          "id": {
            "description": "草稿 ID",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_mirror_state": {
            "default": "",
            "description": "图片镜像状态（M5b）：''=未启用/未镜像；pending=镜像中；mirrored=已转存 COS；failed=失败保持外链",
            "title": "Image Mirror State",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "description": "envelope {draft, source, extensions}；无 api_key 明文",
            "title": "Payload",
            "type": "object"
          },
          "source": {
            "default": "skill",
            "description": "'skill' | 'webui'",
            "title": "Source",
            "type": "string"
          },
          "submission_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "最新一次提交状态（draft_submissions.status）：pending/uploading/published/failed；NULL = 未上架（C1 状态机，T10 采集箱列）",
            "title": "Submission Status"
          },
          "tenant_id": {
            "description": "归属用户（_authenticate_token 的 user_id）",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "更新时间",
            "title": "Updated At"
          },
          "version": {
            "default": 1,
            "description": "乐观并发版本（PATCH 带旧 version，不匹配 → 409）",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "payload"
        ],
        "title": "DraftOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SubmissionTimelineItem": {
        "description": "M2.2: 草稿提交时间线条目（draft_submissions 行，created_at 倒序）。\n\n供 WebUI 展示「这个草稿被提交过几次、到过哪些店、结果如何」。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "提交时间",
            "title": "Created At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "失败/被拒原因",
            "title": "Error Message"
          },
          "extensions": {
            "additionalProperties": true,
            "description": "提交时 extensions 快照（定价/仓库/库存）",
            "title": "Extensions",
            "type": "object"
          },
          "id": {
            "description": "submission 记录 ID（draft_submissions.id）",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "description": "提交状态：pending/uploading/published/failed/rejected（M0.3 写回）",
            "title": "Status",
            "type": "string"
          },
          "store_client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "目标店铺 Ozon Client-Id",
            "title": "Store Client Id"
          },
          "submitted_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "关联任务 ID（ozon_product_tasks.id）",
            "title": "Submitted Task Id"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "SubmissionTimelineItem",
        "type": "object"
      },
      "SubmitResponse": {
        "description": "提交成功响应（含 C5 跨店确认标记）。",
        "examples": [
          {
            "confirm_required": false,
            "draft_id": "a1b2c3d4-0000-4000-8000-000000000001",
            "existing_stores": [],
            "ok": true,
            "status": "pending",
            "submission_id": "a1b2c3d4-0000-4000-8000-000000000002",
            "task_id": "8f1c2c1e-3b7a-4c58-9d2e-1a2b3c4d5e6f"
          }
        ],
        "properties": {
          "confirm_required": {
            "default": false,
            "description": "跨店提醒：该草稿已提交到其他店铺（不硬拦）",
            "title": "Confirm Required",
            "type": "boolean"
          },
          "draft_id": {
            "description": "草稿 ID（多次提交永不变）",
            "format": "uuid",
            "title": "Draft Id",
            "type": "string"
          },
          "existing_stores": {
            "description": "已有提交的店铺 client_id 列表",
            "items": {
              "type": "string"
            },
            "title": "Existing Stores",
            "type": "array"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "status": {
            "default": "pending",
            "description": "提交记录状态：pending/uploading/published/failed",
            "title": "Status",
            "type": "string"
          },
          "submission_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "本次提交记录 ID（draft_submissions.id）",
            "title": "Submission Id"
          },
          "task_id": {
            "default": "",
            "description": "ozon_product_tasks.id",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "draft_id"
        ],
        "title": "SubmitResponse",
        "type": "object"
      },
...
      "TaskDraftResponse": {
        "description": "GET /tasks/{task_id}/draft 响应（M1.1 失败/被拒任务 → 找回采集箱草稿）。\n\n解析顺序：draft_submissions.submitted_task_id → product_task_index.task_id → None\n（直连任务无 submission 行时回落到 product_task_index；都无 → draft_id=None）。",
        "properties": {
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "采集箱草稿 UUID；无关联草稿（直连任务）→ None",
            "title": "Draft Id"
          }
        },
        "title": "TaskDraftResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Ozon 产品上架 Worker — 接收信封、执行 LangGraph 管线、上传 Ozon",
    "title": "Ozon Worker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/drafts/{draft_id}": {
      "delete": {
        "description": "T10 采集箱删除：租户隔离；draft_submissions 由 FK CASCADE 级联删（验收：清空/删除级联删 submissions）。",
        "operationId": "delete_draft_api_v1_drafts__draft_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Draft",
        "tags": [
          "drafts"
        ]
      },
      "get": {
        "operationId": "get_draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "drafts"
        ]
      },
      "patch": {
        "operationId": "patch_draft_api_v1_drafts__draft_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Draft",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/ai/{field}": {
      "post": {
        "description": "单字段 AI 重新生成（T14b）：只读，返回 RU 值，不写回草稿（前端 PATCH 保存）。\n\n错误映射：无/无效 token → 401（_authenticate_token）；未知 field → 400；\n草稿不存在/跨租户 → 404；草稿字段为空或 LLM 失败/含中文拉丁残留 → 422。",
        "operationId": "draft_ai_field_api_v1_drafts__draft_id__ai__field__post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "field",
            "required": true,
            "schema": {
              "title": "Field",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftAiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Ai Field",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/assemble": {
      "post": {
        "description": "一键预组装（v0.70）：LLM 生成整卡上架信息并写回 payload（version++）。\n\n幂等：已含西里尔的字段跳过不重烧 LLM；ozon_attributes 已有内容（跟卖竞品\n属性）不混源。类目建议只写展示键 draft.suggested_category（**绝不写\ndraft.ozon_category**，不劫持管线仲裁链）；三档预估价只写展示键\ndraft.estimated_pricing（RUB；pricing_node 永远按成本+margin 重算）。\n\n错误映射：无/无效 token → 401（_authenticate_token）；草稿不存在/跨租户 → 404。",
        "operationId": "draft_assemble_api_v1_drafts__draft_id__assemble_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "T14b: 单字段 AI 重新生成请求。\n\n请求体仅携带 token（与全站一致：token 在 body 而非 header）；\n草稿读取由 {draft_id} + token 鉴权得到的 tenant_id 共同限定。",
                "properties": {
                  "token": {
                    "description": "mxou API Key（用于 LLM 调用与鉴权）",
                    "title": "Token",
                    "type": "string"
                  }
                },
                "required": [
                  "token"
                ],
                "title": "DraftAiRequest",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftAssembleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Assemble",
        "tags": [
          "drafts"
        ]
      }
    },
...
    "/api/v1/drafts/{draft_id}/resubmit": {
      "post": {
        "description": "失败/被拒草稿重新提交(进行中 → 409)。",
        "operationId": "resubmit_draft_api_v1_drafts__draft_id__resubmit_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resubmit Draft",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/submissions": {
      "get": {
        "description": "M2.2 提交时间线：草稿被提交过几次、到过哪些店、结果如何（created_at 倒序）。\n\n先校验草稿归属（不存在/跨租户 → 404），再返回全部 submission 行。",
        "operationId": "list_submissions_api_v1_drafts__draft_id__submissions_get",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionTimelineItem"
                  },
                  "title": "Response List Submissions Api V1 Drafts  Draft Id  Submissions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Submissions",
        "tags": [
          "drafts"
        ]
      }
    },
    "/api/v1/drafts/{draft_id}/submit": {
      "post": {
        "operationId": "submit_draft_api_v1_drafts__draft_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Draft",
        "tags": [
          "drafts"
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}/draft": {
      "get": {
        "description": "M1.1: task → 采集箱草稿（失败/被拒任务回采集箱改 → 重上）。\n\n租户隔离：task 属于其他租户/不存在 → 404（task_service 内拦截）。\n返回 {\"draft_id\": uuid | None}：直连任务无草稿 → None（前端提示直接重上）。",
        "operationId": "get_task_draft_api_v1_tasks__task_id__draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Draft",
        "tags": [
          "v1",
          "task"
        ]
      }
    },
...
  }
}